refactor: move Cluster testing to test_cluster.py
